updating comments
